Applied BitMask
Jsons and Json now uses BitMask
Arrays and Systems are now final and have private constructor function so no one can obtain their variable.

+++ src/main/java/dev/exefile7f/rheniumcore/util/BitMask.java
@@ -1,21 +1,21 @@
 package dev.exefile7f.rheniumcore.util;
 
 public class BitMask{
-    private long nextBits;
-    private long bits;
+    private int nextBits;
+    private int bits;
     public BitMask(){}
-    public long create(){
+    public int create(){
         if(nextBits >= 64)return 0;
-        return 1L << nextBits++;
+        return 1 << nextBits++;
     }
-    public boolean isSet(long num){
+    public boolean isSet(int num){
         return (bits & num) != 0;
     }
-    public BitMask enable(long num){
+    public BitMask enable(int num){
         bits |= num;
         return this;
     }
-    public BitMask disable(long num){
+    public BitMask disable(int num){
         bits &= ~num;
         return this;
     }
